+++ src/globals/res.js
@@ -1,5 +1,0 @@
-import Vue from 'vue';
-
-Vue.prototype.$res = function (res) {//全局函数1
-    console.log(res);
-};